More work on bringing StarryPy3k up to speed with Giraffe.
At this juncture, the proxy and packet reading is ~working~... but that doesn't say much. There is still a lot of testing to do, and functionality to re-implement.

The main bit is, a user logging in, will correctly have their name and UUID scraped.

IRC connecting also works, though chat messages sent in-channel do not show in-game.